Add go version check to build
Version check is really simple, it checks if REQ_GO_VERSION is in the
captured version string from `go version`. This is an exact match and
has no concept of "at least version x" etc...

Add target dependencies to ensure certain targets, such as
'ensure-version' are always executed for build and dist.