feat: allow specifying config flags via a JSON file as well as env vars
Currently certain variables can be specified as env vars instead of with
command-line flags. This change means that a JSON file can also be used.
This can be enabled by using the `-config-file` flag to point to a JSON
file. The JSON file should use the same key names as the flag names, for
example:

```json
{
	"mode": "full",
	"provider": "vault",
	"vault-namedkey": "xyz"
}
```

I decided to only use this for the `encrypt` and `decrypt` commands as
the `vault-autoconfig` command uses a different flag name for the
encryption key, and I thought it would be confusing for the JSON file to
need two different fields for the same thing.

I implemented this using `urfave/cli`'s `altsrc` package which is
designed for this purpose. It's a little tricky to work out how to use
it, but it ultimately seems reasonable sensible.